feat: create interface of kadis with dashboard and report cases

+++ src/app/home/page.tsx
@@ -29,6 +29,10 @@ export default function page() {
       redirect("/admin/dashboard");
     }
 
+    if (decodeToken.role == "KADIS") {
+      redirect("/head/dashboard");
+    }
+
     localStorage.clear();
     redirect("/auth/login");
   }, []);
